requirements.txt
google-cloud-storage==2.7.0
google-cloud-bigquery==3.10.0
pyspark==3.3.0
python-dotenv==1.0.0
pytest==7.2.0
pandas==1.5.3